+++ CHANGELOG.md
@@ -5,12 +5,17 @@ Change Log
 
 - fixed multi-page queries when `--num N` was greater that the initial result size
 - added `--np` and `--noprompt` options to just search and exit
-- added `-l` `--language` command line options and `language` option in config file
+- added `-l` `--language` command line options and `language` configuration option to set search result language prefernece
 - added `-j` `--first` command line option to open the first result and exit
 - added `--lucky` command line option to open a random result and exit
 - added `week` option to `--time-range` option and enabled `d`, `w`, `m`, `y` as short codes
 - added `--unsafe` command line option as alternative for `--safe-search none`
 - added `--version` command line option
+- switch from using requests to httpx
+- added default http headers set User Agent
+- added `--http-method` command line and `http_method` configuration option to use GET of POST for querys
+- added `--no-verify-ssl` command line and `no-verify-ssl` configuraiton option for sites with self signed ccertificates
+- added `--noua` command line option to disable User Agent
 
 ## v0.1.0